refactor: rename the modelmigration/v1_28 package to v28
The next release is 28.0.0, not 1.28.0. Older packages keep their v1_N
names, those releases really were 1.x.
